Simplify the documentation of `reverse_geocode`
